+++ packages/notification_transport/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.0.1
 
+* docs: update readme.
+* ci: fix integration tests and example.
 * ci: separate darwin integration tests.
 
 ## 2.0.0